@@ -4,21 +4,21 @@ import com.google.common.collect.Multimap;
 import com.google.common.collect.Sets;
 import com.thevoxelbox.voxelsniper.brush.IBrush;
 import org.bukkit.util.Vector;
-import org.junit.Assert;
-import org.junit.Before;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.TestInstance;
+import org.junit.jupiter.api.TestInstance.Lifecycle;
 import org.mockito.Mockito;
 
 import java.util.Set;
 
-/**
- *
- */
+@TestInstance(Lifecycle.PER_CLASS)
 public class BrushesTest
 {
     private Brushes brushes;
 
-    @Before
+    @BeforeEach
     public void setUp() throws Exception
     {
         brushes = new Brushes();
@@ -36,27 +36,27 @@ public class BrushesTest
     {
         IBrush brush = Mockito.mock(IBrush.class);
         brushes.registerSniperBrush(brush.getClass(), "mockhandle", "testhandle");
-        Assert.assertEquals(brush.getClass(), brushes.getBrushForHandle("mockhandle"));
-        Assert.assertEquals(brush.getClass(), brushes.getBrushForHandle("testhandle"));
-        Assert.assertNull(brushes.getBrushForHandle("notExistant"));
+        Assertions.assertEquals(brush.getClass(), brushes.getBrushForHandle("mockhandle"));
+        Assertions.assertEquals(brush.getClass(), brushes.getBrushForHandle("testhandle"));
+        Assertions.assertNull(brushes.getBrushForHandle("notExistant"));
     }
 
     @Test
     public void testRegisteredSniperBrushes() throws Exception
     {
-        Assert.assertEquals(0, brushes.registeredSniperBrushes());
+        Assertions.assertEquals(0, brushes.registeredSniperBrushes());
         IBrush brush = Mockito.mock(IBrush.class);
         brushes.registerSniperBrush(brush.getClass(), "mockhandle", "testhandle");
-        Assert.assertEquals(1, brushes.registeredSniperBrushes());
+        Assertions.assertEquals(1, brushes.registeredSniperBrushes());
     }
 
     @Test
     public void testRegisteredSniperBrushHandles() throws Exception
     {
-        Assert.assertEquals(0, brushes.registeredSniperBrushHandles());
+        Assertions.assertEquals(0, brushes.registeredSniperBrushHandles());
         IBrush brush = Mockito.mock(IBrush.class);
         brushes.registerSniperBrush(brush.getClass(), "mockhandle", "testhandle");
-        Assert.assertEquals(2, brushes.registeredSniperBrushHandles());
+        Assertions.assertEquals(2, brushes.registeredSniperBrushHandles());
     }
 
     @Test
@@ -65,9 +65,9 @@ public class BrushesTest
         IBrush brush = Mockito.mock(IBrush.class);
         brushes.registerSniperBrush(brush.getClass(), "mockhandle", "testhandle");
         Set<String> sniperBrushHandles = brushes.getSniperBrushHandles(brush.getClass());
-        Assert.assertTrue(sniperBrushHandles.contains("mockhandle"));
-        Assert.assertTrue(sniperBrushHandles.contains("testhandle"));
-        Assert.assertFalse(sniperBrushHandles.contains("notInSet"));
+        Assertions.assertTrue(sniperBrushHandles.contains("mockhandle"));
+        Assertions.assertTrue(sniperBrushHandles.contains("testhandle"));
+        Assertions.assertFalse(sniperBrushHandles.contains("notInSet"));
     }
 
     @Test
@@ -76,10 +76,10 @@ public class BrushesTest
         IBrush brush = Mockito.mock(IBrush.class);
         brushes.registerSniperBrush(brush.getClass(), "mockhandle", "testhandle");
         Multimap<Class<? extends IBrush>,String> registeredBrushesMultimap = brushes.getRegisteredBrushesMultimap();
-        Assert.assertTrue(registeredBrushesMultimap.containsKey(brush.getClass()));
-        Assert.assertFalse(registeredBrushesMultimap.containsKey(IBrush.class));
-        Assert.assertTrue(registeredBrushesMultimap.containsEntry(brush.getClass(), "mockhandle"));
-        Assert.assertTrue(registeredBrushesMultimap.containsEntry(brush.getClass(), "testhandle"));
-        Assert.assertFalse(registeredBrushesMultimap.containsEntry(brush.getClass(), "notAnEntry"));
+        Assertions.assertTrue(registeredBrushesMultimap.containsKey(brush.getClass()));
+        Assertions.assertFalse(registeredBrushesMultimap.containsKey(IBrush.class));
+        Assertions.assertTrue(registeredBrushesMultimap.containsEntry(brush.getClass(), "mockhandle"));
+        Assertions.assertTrue(registeredBrushesMultimap.containsEntry(brush.getClass(), "testhandle"));
+        Assertions.assertFalse(registeredBrushesMultimap.containsEntry(brush.getClass(), "notAnEntry"));
     }
 }
