@@ -2,16 +2,20 @@ package com.thevoxelbox.voxelsniper.util;
 
 import com.google.common.collect.ImmutableSet;
 import org.bukkit.Keyed;
+import org.bukkit.NamespacedKey;
 import org.bukkit.Tag;
 
 import java.util.Set;
+import org.jetbrains.annotations.NotNull;
 
 public class SetTag<T extends Keyed> implements Tag<T> {
 
     private final Set<T> values;
+    private final NamespacedKey key;
 
-    public SetTag(Set<T> values) {
+    public SetTag(Set<T> values, NamespacedKey key) {
         this.values = ImmutableSet.copyOf(values);
+        this.key = key;
     }
 
     @Override
@@ -22,5 +26,10 @@ public class SetTag<T extends Keyed> implements Tag<T> {
     @Override
     public Set<T> getValues() {
         return values;
+    }
+
+    @Override
+    public @NotNull NamespacedKey getKey() {
+        return key;
     }
 }
